requirements.txt
# RLM vs RAG Research - Dependencies

# LLM & API
openai>=1.0.0
google-generativeai>=0.8.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# RAG - Embeddings & Vector Store
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Data & Evaluation
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0  # HuggingFace datasets for OOLONG

# Excel support (for data layout benchmark)
openpyxl>=3.1.0

# Utilities
rich>=13.0.0
tqdm>=4.65.0
pyyaml>=6.0
